代码逻辑优化

+++ src/com/lls/sms/ui/dialog/DeleteStudentDialog.java
@@ -17,7 +17,7 @@ import java.awt.event.ActionListener;
  * @Date: 2023/6/2 14:24
  */
 
-public class DeleteStudentDialog extends JDialog{
+public class DeleteStudentDialog extends JDialog {
     private JPanel jPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 10, 20));
     private JLabel idLabel = new JLabel("学号：");
     private JTextField idText = new JTextField();
@@ -40,21 +40,22 @@ public class DeleteStudentDialog extends JDialog{
         container.add(jPanel);
 
 
-
         deleteBtn.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
                 int response = JOptionPane.showConfirmDialog(null, "你确定要删除这个学生🐎？", "确认操作", JOptionPane.YES_NO_OPTION);
-                if(response==0){
-                    studentService.deleteStudent(Long.valueOf(idText.getText()));
-                    JOptionPane.showMessageDialog(null, "删除成功", "操作提示", JOptionPane.PLAIN_MESSAGE);
-                }else if(response==1){
+                if (response == 0) {
+                    if (studentService.getStudent(Long.valueOf(idText.getText())).getName() != null) {
+                        studentService.deleteStudent(Long.valueOf(idText.getText()));
+                        JOptionPane.showMessageDialog(null, "删除成功", "操作提示", JOptionPane.PLAIN_MESSAGE);
+                    } else {
+                        JOptionPane.showMessageDialog(null, "没有该学号的学生", "操作异常", JOptionPane.ERROR_MESSAGE);
+                    }
+                } else if (response == 1) {
                     JOptionPane.showMessageDialog(null, "取消删除", "操作提示", JOptionPane.PLAIN_MESSAGE);
                 }
             }
         });
-
-
 
 
         setSize(350, 450);
